Rename TraceBuffer -> TraceBufferV1, extract interface (#3091)
Extract a common interface for TraceBufferV1 and V2.
Renames TraceBuffer to TraceBufferV1, in preparation for TBV2

Bug: b/447426810
