+++ Android.bp
@@ -16557,7 +16557,7 @@ filegroup {
         "src/tracing/service/metatrace_writer.cc",
         "src/tracing/service/packet_stream_validator.cc",
         "src/tracing/service/random.cc",
-        "src/tracing/service/trace_buffer.cc",
+        "src/tracing/service/trace_buffer_v1.cc",
         "src/tracing/service/tracing_service_impl.cc",
     ],
 }
@@ -16568,7 +16568,7 @@ filegroup {
     srcs: [
         "src/tracing/service/histogram_unittest.cc",
         "src/tracing/service/packet_stream_validator_unittest.cc",
-        "src/tracing/service/trace_buffer_unittest.cc",
+        "src/tracing/service/trace_buffer_v1_unittest.cc",
         "src/tracing/service/tracing_service_impl_unittest.cc",
         "src/tracing/service/zlib_compressor_unittest.cc",
     ],
