+++ src/tracing/service/trace_buffer_v1_unittest.cc
@@ -28,7 +28,7 @@
 #include "perfetto/ext/tracing/core/trace_packet.h"
 #include "perfetto/protozero/proto_utils.h"
 #include "src/base/test/vm_test_utils.h"
-#include "src/tracing/service/trace_buffer.h"
+#include "src/tracing/service/trace_buffer_v1.h"
 #include "src/tracing/test/fake_packet.h"
 #include "test/gtest_and_gmock.h"
 
@@ -40,9 +40,9 @@ using ::testing::IsEmpty;
 
 class TraceBufferTest : public testing::Test {
  public:
-  using SequenceIterator = TraceBuffer::SequenceIterator;
-  using ChunkMetaKey = TraceBuffer::ChunkMeta::Key;
-  using ChunkRecord = TraceBuffer::ChunkRecord;
+  using SequenceIterator = TraceBufferV1::SequenceIterator;
+  using ChunkMetaKey = TraceBufferV1::ChunkMeta::Key;
+  using ChunkRecord = TraceBufferV1::ChunkRecord;
 
   static constexpr uint8_t kContFromPrevChunk =
       SharedMemoryABI::ChunkHeader::kFirstPacketContinuesFromPrevChunk;
@@ -57,10 +57,10 @@ class TraceBufferTest : public testing::Test {
     if (trace_buffer_) {
       const size_t used_size = trace_buffer_->used_size();
       ASSERT_LE(used_size, trace_buffer_->size());
-      trace_buffer_->EnsureCommitted(trace_buffer_->size());
+      trace_buffer()->EnsureCommitted(trace_buffer_->size());
       bool zero_padded = true;
       for (size_t i = used_size; i < trace_buffer_->size(); ++i) {
-        bool is_zero = static_cast<char*>(trace_buffer_->data_.Get())[i] == 0;
+        bool is_zero = static_cast<char*>(trace_buffer()->data_.Get())[i] == 0;
         zero_padded = zero_padded && is_zero;
       }
       ASSERT_TRUE(zero_padded);
@@ -68,13 +68,13 @@ class TraceBufferTest : public testing::Test {
   }
 
   FakeChunk CreateChunk(ProducerID p, WriterID w, ChunkID c) {
-    return FakeChunk(trace_buffer_.get(), p, w, c);
+    return FakeChunk(trace_buffer(), p, w, c);
   }
 
   void ResetBuffer(
       size_t size_,
       TraceBuffer::OverwritePolicy policy = TraceBuffer::kOverwrite) {
-    trace_buffer_ = TraceBuffer::Create(size_, policy);
+    trace_buffer_ = TraceBufferV1::Create(size_, policy);
     ASSERT_TRUE(trace_buffer_);
   }
 
@@ -145,26 +145,31 @@ class TraceBufferTest : public testing::Test {
   }
 
   SequenceIterator GetReadIterForSequence(ProducerID p, WriterID w) {
-    TraceBuffer::ChunkMeta::Key key(p, w, 0);
-    return trace_buffer_->GetReadIterForSequence(
-        trace_buffer_->index_.lower_bound(key));
+    TraceBufferV1::ChunkMeta::Key key(p, w, 0);
+    return trace_buffer()->GetReadIterForSequence(
+        trace_buffer()->index_.lower_bound(key));
   }
 
   void SuppressClientDchecksForTesting() {
-    trace_buffer_->suppress_client_dchecks_for_testing_ = true;
+    trace_buffer()->suppress_client_dchecks_for_testing_ = true;
   }
 
   std::vector<ChunkMetaKey> GetIndex() {
     std::vector<ChunkMetaKey> keys;
-    keys.reserve(trace_buffer_->index_.size());
-    for (const auto& it : trace_buffer_->index_)
+    keys.reserve(trace_buffer()->index_.size());
+    for (const auto& it : trace_buffer()->index_)
       keys.push_back(it.first);
     return keys;
   }
 
-  uint8_t* GetBufData(const TraceBuffer& buf) { return buf.begin(); }
-  TraceBuffer* trace_buffer() { return trace_buffer_.get(); }
-  size_t size_to_end() { return trace_buffer_->size_to_end(); }
+  uint8_t* GetBufData(const TraceBuffer& buf) {
+    return static_cast<const TraceBufferV1&>(buf).begin();
+  }
+
+  TraceBufferV1* trace_buffer() {
+    return static_cast<TraceBufferV1*>(trace_buffer_.get());
+  }
+  size_t size_to_end() { return trace_buffer()->size_to_end(); }
 
  private:
   std::unique_ptr<TraceBuffer> trace_buffer_;
